+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased][unreleased]
+- Do not raise for expression that are not supported, instead return `{:error, reason}`.
+- Using index on filter expression in data type that is not a list, result in empty list, ex: `Warpath.query!([%{}, :a, "b"], "$[?(@[0] > 1)]") => []`.
 
 ## [0.3.0-rc.2] - 2020-05-25
 - Improve performance remove overhead caused by Task.async_stream and Stream module.